+++ motor/src/Motor/FSM/Class.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE ConstraintKinds        #-}
 {-# LANGUAGE DataKinds              #-}
 {-# LANGUAGE FlexibleContexts       #-}
+{-# LANGUAGE FlexibleInstances      #-}
 {-# LANGUAGE GADTs                  #-}
 {-# LANGUAGE MultiParamTypeClasses  #-}
 {-# LANGUAGE PolyKinds              #-}
@@ -20,6 +21,7 @@ module Motor.FSM.Class
 
 import           Control.Monad.Indexed
 import           Data.Row.Records
+import           GHC.OverloadedLabels
 import           GHC.TypeLits          (Symbol)
 
 -- * FSM monad
@@ -27,16 +29,28 @@ import           GHC.TypeLits          (Symbol)
 -- | An indexed monad for finite-state machines, managing the state
 -- of named resources.
 class IxMonad m => MonadFSM (m :: (Row *) -> (Row *) -> * -> *) where
-  -- | Creates a new resource and returns its 'Name'.
+  -- | Creates a new resource by name.
   new :: Name n -> a -> m r (Extend n a r) ()
+  -- | Returns an existing resource.
+  get :: HasType n a r => Name n -> m r r a
   -- | Deletes an existing resource named by its 'Name'.
   delete :: Name n -> m r (r .- n) ()
   -- | Replaces the state of an existing resource named by its 'Name'.
-  enter :: Name n -> b -> m r (Modify n b r) ()
-  -- | Run another 'MonadFSM' computation, with empty resource rows,
+  enter ::
+       (HasType n a r, Modify n b r ~ r')
+    => Name n
+    -> b
+    -> m r r' ()
+  -- | Updates the state, using a pure function, of an existing
+  -- resource named by its 'Name'.
+  update :: (HasType n a r, Modify n a r ~ r) => Name n -> (a -> a) -> m r r ()
+  -- | Embed another 'MonadFSM' computation, with empty resource rows,
   -- in this computation.
   call :: m Empty Empty a -> m r r a
 
 -- | A name of a resource, represented using a 'Symbol'.
 data Name (n :: Symbol) where
   Name :: KnownSymbol n => Name n
+
+instance (KnownSymbol n, n ~ n') => IsLabel n (Name n') where
+  fromLabel = Name :: Name n'