+++ motor-reflection/CHANGELOG.md
@@ -1,5 +1,13 @@
-* ?
-  - Use row-types instead of CTRex
+* 0.3.0
+  - Use row-types instead of CTRex.
+  - Rename `:->` operator to `:=`.
+  - Provide better default constraints from using `Actions`.
+  - Make `Actions` a data type and use it promoted.
+  - Add `Remain` action.
+  - Add `get` operator, returning the current state of a named
+    resource.
+  - Add `update` operator, mapping an update function over the current
+    state of a named resource.
 * 0.2.0.0
   - Add `motor-reflection` and `motor-diagrams` packages.
 * 0.1.1.0